DCA11Y-1145: Add handling of loosely-defined major versions
