+++ frontend-maven-plugin/src/main/java/com/github/eirslett/maven/plugins/frontend/mojo/InstallNodeAndYarnMojo.java
@@ -2,7 +2,7 @@ package com.github.eirslett.maven.plugins.frontend.mojo;
 
 import com.github.eirslett.maven.plugins.frontend.lib.FrontendPluginFactory;
 import com.github.eirslett.maven.plugins.frontend.lib.NodeVersionDetector;
-import com.github.eirslett.maven.plugins.frontend.lib.NodeVersionParser;
+import com.github.eirslett.maven.plugins.frontend.lib.NodeVersionHelper;
 import com.github.eirslett.maven.plugins.frontend.lib.ProxyConfig;
 import com.github.eirslett.maven.plugins.frontend.lib.YarnInstaller;
 import org.apache.maven.execution.MavenSession;
@@ -14,7 +14,7 @@ import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.settings.Server;
 import org.apache.maven.settings.crypto.SettingsDecrypter;
 
-import static com.github.eirslett.maven.plugins.frontend.lib.NodeVersionParser.fixupMinorVersionErrors;
+import static com.github.eirslett.maven.plugins.frontend.lib.NodeVersionHelper.getDownloadableVersion;
 import static com.github.eirslett.maven.plugins.frontend.mojo.YarnUtils.isYarnrcYamlFilePresent;
 import static java.util.Objects.isNull;
 
@@ -89,11 +89,11 @@ public final class InstallNodeAndYarnMojo extends AbstractFrontendMojo {
             throw new LifecycleExecutionException("Node version could not be detected from a file and was not set");
         }
 
-        if (!NodeVersionParser.validateVersion(nodeVersion)) {
+        if (!NodeVersionHelper.validateVersion(nodeVersion)) {
             throw new LifecycleExecutionException("Node version (" + nodeVersion + ") is not valid. If you think it actually is, raise an issue");
         }
 
-        String validNodeVersion = fixupMinorVersionErrors(nodeVersion);
+        String validNodeVersion = getDownloadableVersion(nodeVersion);
 
         boolean isYarnYamlFilePresent = isYarnrcYamlFilePresent(this.session, this.workingDirectory);
